Move file in a new architecture not oriented around cutting it down Add periodogram method More static dispatch
