+++ lib/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+add_subdirectory(complex)
+add_subdirectory(dft)
+add_subdirectory(plot)
+add_subdirectory(report)
