+++ lib/complex/CMakeLists.txt
@@ -1,0 +1,7 @@
+
+
+add_library(Complex INTERFACE)
+
+target_include_directories(Complex INTERFACE 
+                       $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include> # for headers when building
+                       $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
