+++ lib/report/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+add_library(Report INTERFACE)
+
+target_include_directories(Report INTERFACE 
+                       $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include> # for headers when building
+                       $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
