--- spp/CMakeLists.txt
@@ -1,4 +1,0 @@
-
-add_subdirectory(dft)
-add_subdirectory(plot)
-add_subdirectory(tools)
