--- spp/dft/include/dft/dft_policies.hpp
@@ -1,23 +1,0 @@
-
-#pragma once
-
-/**
- * @defgroup dft_policy
- * @addtogroup dft
- */
-namespace dft
-{
-namespace policy
-{
-/**
- * @brief fftw_t - fftw policy for selecting fftw for computing dft
- * @addtogroup dft_policy
- */
-struct Fftw{};
-
-inline constexpr Fftw fftw{};
-
-inline constexpr auto defaut = fftw;
-
-}
-}
