--- spp/dft/src/fftw3/fftw3_wrapper.cpp
@@ -1,6 +1,0 @@
-#include "dft/fftw3/fftw3_wrapper.hpp"
-
-namespace dft
-{
-
-}
