--- spp/tools/CMakeLists.txt
@@ -1,2 +1,0 @@
-
-add_subdirectory(pico)
