--- spp/tools/pico/CMakeLists.txt
@@ -1,6 +1,0 @@
-
-add_executable(pico 
-   src/main.cpp
-)
-
-target_link_libraries(pico PRIVATE dft CLI11::CLI11)
