+++ tools/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+foreach(TOOL ${TOOLS})
+   add_subdirectory(${TOOL}) 
+endforeach()
