+++ tools/pico/CMakeLists.txt
@@ -1,0 +1,6 @@
+
+add_executable(Pico 
+   src/main.cpp
+)
+
+target_link_libraries(Pico PRIVATE Dft CLI11::CLI11)
